+++ server/src/main/kotlin/tk/jomp16/habbo/game/user/subscription/HabboSubscription.kt
@@ -71,16 +71,13 @@ class HabboSubscription(private val habboSession: HabboSession) {
             months = ChronoUnit.MONTHS.between(currentTime, subscription?.expire).toInt()
             elapsedDays = ChronoUnit.DAYS.between(subscription?.activated, currentTime).toInt()
             minutes = ChronoUnit.MINUTES.between(currentTime, subscription?.expire).toInt()
-            days = Period.between(LocalDate.now(), LocalDate.now().plusDays(days.toLong()).minusMonths(months.toLong())).days.toInt()
+            days = Period.between(LocalDate.now(), LocalDate.now().plusDays(days.toLong()).minusMonths(months.toLong())).days
 
             if (days == 0) days = 1
         }
 
         habboSession.sendHabboResponse(Outgoing.SUBSCRIPTION_STATUS, HabboSubscription.CLUB_TYPE, active, days, if (months >= 1) months - 1 else months, elapsedDays, minutes)
-        habboSession.sendHabboResponse(Outgoing.USER_RIGHTS, if (habboSession.userInformation.vip || habboSession.habboSubscription.validUserSubscription) 2 else 0,
-                habboSession.userInformation.rank,
-                habboSession.userInformation.ambassador
-        )
+        habboSession.sendHabboResponse(Outgoing.USER_RIGHTS, if (habboSession.userInformation.vip || habboSession.habboSubscription.validUserSubscription) 2 else 0, habboSession.userInformation.rank, habboSession.userInformation.ambassador)
     }
 
     companion object {
